refactor(pr): PR 详情工作区从 MainPane 剥离到 features/pr/
layout/MainPane 原含整个 PR 评审工作区（头部 / tab 栏 / tab 内容 + 全部 PR
详情状态），与「布局壳」职责混淆。剥离：
- features/pr/PrPanel —— PR 工作区容器：tab / diff 视图选项 / 评论 + 提交计数 /
  草稿池 / 发布弹窗状态 + tab 内容装配
- features/pr/PrHeader —— 标题 / 元信息 + 动作区（浏览器 / 提交评论 / 合并 / 审批）
- features/pr/PrTabs —— tab 栏 + diff 工具条（空白 / blame / 并排-内联）
- layout/MainPane 退化为薄壳（503→96 行）：仅主区 + 「未选 PR」空态，选中时挂 PrPanel

对外 props 接口不变，纯结构剥离，无逻辑/界面变更。
